Creates a new cohort_data_definition object

This function is used to create a new cohort_data_definition object. Results from our audit take two forms - either we want to collect aggregated data from a temporally grouped cohort of patients; or we want access to a subset of columns from that cohort. Both the cohorts and the data that we want tend to have a many-to-many relationship with each other - ie. we often want the same aggregated results from multiple different cohort types (in the case of SSNAP team and patient-centred results). In other cases we want to apply different rules to the same cohorts in different circumstances (in the case of SSNAP just Key Indicators or just audit compliance in some cases, whereas on others the whole portfolio is needed). Using cohort_data_definitions we can combine all the permutations we are likely to use commonly; or make our own custom permutations eg. when creating user filters in a realtime dashboard.

cohort_data_definition(cohort_definition, audit_output_spec)

Arguments

cohort_definition
A cohort definition object, built using the cohort_definition function.

audit_output_spec
An output definition, built using the create_output function.

Value

The created cohort_definition object.