Create cohorts of SSNAP data

The cohorts command takes source data from the CSV file and returns a range of different cohorts needed to produce reports or undertake analysis. The range of cohorts that can be returned is described below.

The idea is to request all the cohorts you want in a single pass at the beginning of the analysis process. Some cohorts are dependent upon other cohorts: this function allows us to efficiently work out which cohorts need to be kept, and which can be thrown away after using them to calculate our end products, saving time and memory versus saving all the cohorts (potentially wasting vast amounts of memory) or throwing them all away (forcing multiple passes over the raw data which is computationally expensive).

cohorts(ssnap_data, cohorts_requested, period_type = c("months", "periods",
  "financialyears"), last_period, number_of_periods,
  group_aggr_cohorts_by = c("team", "national"), verbose = FALSE)

Arguments

ssnap_data
This may be either:

The filename of a SSNAP CSV file (NB currently only CSV files are supported, not encrypted ZIP files).

A tibble (a loaded data frame created by the ssnapcsv function from the SSNAPInterface package).

cohorts_requested
A vector of strings defining which cohorts you wish to return. These cohorts can be any or all of the following:

raw72hrs Returns all the records that will be included in the 72hr cohort, filtered by the specified date range. Only inpatient records are included (any team beginning with 'C' or 'c' is rejected). Columns included:

ProClinV1Id and PatientId - the anonymised record and patient identifiers used for linking records / cohorts

TeamCode - The team caring for the patient.

EthnicityCatUK - Ethnicity types in name format, UK-wide (so that clashing codes between NI and England are differentiated).

CreatedDateTime (the date and time the record was created)

All columns beginning with S1-S3.

LockedS1, S2 and S3 (the lock status for each section)

lockedTo72hrs Returns all members of the raw72hrs cohort that are recorded as locked to 72hrs. Note that if a patient moves between teams in the first 72hrs their opening record may not be locked to 72hrs and the subsequent team may lock the record at 72hrs. All columns from raw72hrs are returned but the records filtered as above.

firstTeamLockedTo72hrs Returns the first team in a patient's episode that locks the record to 72hrs. We use this to calculate 72hr measures. NB this isn't guaranteed to be the first team that treats the patient. It uses lockedto72hrs as the input cohort. All columns are preserved.

firstTeamAdmittingPatient Returns the team which first admitted the patient. All 72hr measures are attributed to the team which opened the record. It uses raw72hrs to find the earliest record ID for each patient.

team72hrKIs Takes the first team which locked the record to 72hrs (firstTeamLockedTo72hrs), and calculates the key indicators for the team 72hr cohort. We then take these plus the patient ID, and we apply them to the admitting team.

patient72hrKIs Takes the first team which locked the record to 72hrs (firstTeamLockedTo72hrs), and calculates the key indicators for the patient 72hr cohort. We then take these plus the patient ID, and we apply them to all teams treating the patient.

rawPost72hrs Returns all the records that will be included in post-72hr cohorts, filtered by the specified date range. Only inpatient records are included (any team beginning with 'C' is rejected). Columns included:

ProClinV1Id and PatientId - the anonymised record and patient identifiers used for linking records / cohorts

TeamCode - The team caring for the patient.

EthnicityCatUK - Ethnicity types in name format, UK-wide (so that clashing codes between NI and England are differentiated).

CreatedDateTime (the date and time the record was created)

All columns beginning with S4-S7.

LockedS1 - LockedS7 (the lock status for each section)

LockedS1DateTime - LockedS7DateTime (the date and time of locking for each section - this is used to calculate the 'record lock time' for audit compliance.)

TransferActionDateTime - the date and time the record was transferred (used for audit compliance).

teamDischargeKIs Applies the aggregation functions for the SSNAP KI and Audit Compliance measures for the Team post-72hr (discharge) cohort. This will return things like monthly clock start to scan times etc, grouped by the TeamCode and period_type parameter below.

teamPost72hrAllTeamsCohortKIs This takes all patients who are locked to Sections 4 and 7, OR are transferred to another team, and applies a number of calculated fields to the cohort. These are all marked 'TCKI' for Key Indicators, or 'AC' for calculated fields which affect audit compliance.

teamPost72hrAllTeamsAggrKIs This takes teamPost72hrAllTeamsCohortKIs as an input, and applies the aggregation functions for the Team post-72hr cohort.

teamAt7Days This finds either the current team that the patient is in at 7 days, or the last patient that the team was in (if they were discharged or died within 7 days).

team7DayCohortKIs This uses the teamAt7Days cohort, and applies the information gathered on the patient at discharge (to ensure all fields are up to date). A series of calculated fields are then created based upon this data.

patientPost72hrCohortKIs This fetches all the patients involved in the patient's care. NB some of those may fall outside the search date so we have to link to the raw SSNAP data to fetch all the records associated with the episode.

ssnapScores This takes the aggregated value cohorts and converts KI measures into SSNAP KI scores and Domain scores.

period_type
A string determining what the gap between time periods should be. The factor should either be 'months', 'periods' or financialyears'. Months and financialyears are self-explanatory. Periods represent reporting periods - these were initially quarterly, then were trimesterly, then quarterly again. Periods are referred to by their numeric number, rather than a defined time period to avoid confusion in the calcalations.

last_period
The last period. The type of this value will depend upon the value of period_type If the period_type is 'months' then last_period should be a Date type (the day will be ignored). If period_type is periods' then this should be a numeric value starting at 1 (for Q1 2013). If period_type is 'years' then this should be a numeric value representing the year at the start of that financial year, eg. '2013' represents the 2013-4 financial year.

number_of_periods
The number of periods to be fetched in the report. Common requests for existing reports are 12 months, or the last 4 reporting periods, although for specific requests different output might be requested.

group_aggr_cohorts_by
If any of the cohorts supplied use aggregated data, then this parameter is used to choose the grouping level. Suitable values are 'team' (default) or 'national'.

verbose
The verbose option allows you to see which cohorts are produced and which fields requested from each cohorts. This is intended for developing cohorts and measures, or for optimisation work.

Value

A named list of data frames containing the requested cohorts.