Create outputs from an audit_measure object

create_output is a method to produce a set of instructions to pass to dplyr::summarise, which is then applied over a cohort to produce our cohort results. The instructions are passed as rlang::quos, which can be appended to the summary step using the rlang::!!! operator. Output columns have a consistent naming convention.

# S3 method for audit_measure
create_output(x, numerator = NULL,
  output_type = c("median", "quartiles", "d_n_pct", "d_n_percentage",
  "pct", "percentage"), ...)

Arguments

x
An audit_measure object.

numerator
If the measure has more than one numerator and only one of the numerators is required, this should be set to the name of the numerator. If only one numerator is specified in the measure, or all numerators are to be used, then this should be omitted (or set to NULL).

output_type
Determines which sort of aggregated results are being requested from the data:

median Only applicable to audit_measures whose measure_type is set to continuous. It calculates the median and returns it in a column named with the audit_measure's stem name followed by Q2, e.g. myMeasureQ2. Multiple numerators are not supported by the median output type and will generate an error. This may be fixed if there is a need for it in future. Exclusion criteria are also not currently handled - this is likely to be addressed in future.

quartiles This option is similar to median but in addition to generating Q2, it also creates the lower and upper quartiles, suffixed by Q1 and Q3.

...
Other parameters which may be defined according to the class.