internal_bank_holidays_ni

internal_bank_holidays_ni is a list of bank holidays in Northern Ireland taken from https://www.gov.uk/bank-holidays. NB these are England and Wales bank holidays plus St Patrick's Day and Battle of the Boyne.

internal_bank_holidays_ni

Format

An object of class Date of length 81.

Details

These are needed to analyse differences between normal working hours, weekends and public holidays.