Create a measure looking at proportions of something

Many of the measures we use focus on denominators, numerators or percentages, or breaking down a field into absolute numbers or percentages of possible values or ranges. The function ensures a consistent naming structure and readable code whilst avoiding repetition.

measure_proportion(stem_name, population, exclusions = NULL, measure)

Arguments

stem_name
The name for our new measure. This will form the 'stem' name (ie. suffixes will be added to it for the various components).

population
A function to calculate the COUNT of patients in the population. This is very likely to be deprecated as it should always be length of the number of valid entries in TeamCode or similar. The population will be stored as the string provided in indicator_name suffixed with D (for denominator) and will consist of the length calculation minus the number of excluded patients from the equation given in exclusions.

exclusions
A quosure describing exclusion criteria. These are applied to the population and to all the numerators. By default no exclusions apply.

measure
Either a quosure describing how to calculate the numerator, OR a named list of quosures describing various groups of numerators. The name in the list is used as a suffix to the stem_name. Each numerator is then further suffixed by N.